Reorganize frontend source hierarchy
