+++ app/frontend/src/app/content-tabs/ContentTabs.tsx
@@ -5,8 +5,8 @@
 import * as React from "react";
 import { IModelApp, IModelConnection } from "@bentley/imodeljs-frontend";
 import { Ruleset } from "@bentley/presentation-common";
-import { TabView, TabViewItem } from "../ui-framework/TabView/TabView";
-import { appLayoutContext } from "./AppContext";
+import { TabView, TabViewItem } from "../../ui-framework/TabView/TabView";
+import { appLayoutContext } from "../AppContext";
 import { Editor } from "./editor/Editor";
 import { Viewport } from "./viewport/Viewport";
