+++ app/frontend/src/app/content-tabs/editor/Editor.scss
