+++ app/frontend/src/app/content-tabs/viewport/Viewport.tsx
@@ -7,7 +7,7 @@ import { Id64String } from "@bentley/bentleyjs-core";
 import { IModelConnection } from "@bentley/imodeljs-frontend";
 import { viewWithUnifiedSelection } from "@bentley/presentation-components";
 import { ViewportComponent } from "@bentley/ui-components";
-import { backendApiContext } from "../AppContext";
+import { backendApiContext } from "../../AppContext";
 
 export interface ViewportProps {
   imodel: IModelConnection;
