+++ app/frontend/src/app/utils/VerticalStack.scss
