+++ app/frontend/src/ui-framework/MemoryUISettingsStorage.ts
